Move sc2d test to tests/ directory
